backend/requirements.txt
# LifeFlow Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dateutil==2.8.2

# 生产环境需要 PostgreSQL 驱动，开发环境不需要
# psycopg2-binary==2.9.9
